クエリパラメータ処理をカスタムフックに分離
