+++ frontend/src/hooks/index.js
@@ -2,3 +2,4 @@ export { default as useCryptidCount } from "./useCryptidCount";
 export { default as useLatestCryptids } from "./useLatestCryptids";
 export { default as useCryptids } from "./useCryptids";
 export { default as useHandleSearch } from "./useHandleSearch";
+export { default as useCryptidQueryParams } from "./useCryptidQueryParams";
